Make install-dependencies.sh more generic
This script can now run on multiple Linux distributions, including
Ubuntu, Debian, CentOS and Rocky Linux. Moreover:

 - it now supports two options:
    -u  install the libraries required for unit tests
    -d  install doxygen

 - the download of the library sources is skipped if the tarballs are
   already there.

+++ ci-tests/install-dependencies.sh
@@ -3,13 +3,30 @@
 # install-dependencies.sh: Install the dependencies required for
 # building FeeLLGood.
 #
-# This is a support script for the automated tests performed by GitHub
-# Actions. It is not intended for end-users, and it contains a few calls
-# to `sudo'. Use it at your own risk.
+# This is a support script for automated CI tests. It is not intended
+# for end-users, and it contains a few calls to `sudo'. Use it at your
+# own risk.
+#
+# Options:
+#   -u  install the libraries required for unit tests
+#   -d  install doxygen
 
 ########################################################################
 # The following few commands are not part of the documented installation
 # procedure.
+
+# Parse options.
+while getopts "ud" OPT; do
+    case "${OPT}" in
+        u) unit_tests="true";;
+        d) doxygen="true";;
+        ?) exit 1;;
+    esac
+done
+
+# Identify the OS.
+eval $(grep -E '^(ID|PRETTY_NAME)=' /etc/os-release)
+echo -e "\n*** Building on $PRETTY_NAME"
 
 # Set shell options:
 # -e: exit as soon as a command fails
@@ -25,15 +42,40 @@ job_count=$(getconf _NPROCESSORS_ONLN)
 # apt-get: -y, make: -j, wget: -nv, unzip: -q
 
 # Install required packages.
-sudo apt-get update -q
-sudo apt-get install -y libboost-dev \
-    libboost-system-dev libboost-filesystem-dev libboost-test-dev
+packages="unzip make cmake git"
+if [[ "$ID" =~ centos|rocky ]]; then
+    packages="$packages wget gcc-c++ boost-devel python3"
+    sudo dnf check-update -q || true
+    sudo dnf upgrade -y libarchive
+    if [ "$unit_tests" = "true" ]; then
+        : # No extra packages required for unit testing
+    fi
+    if [ "$doxygen" = "true" ]; then
+        sudo dnf config-manager --set-enabled powertools
+        packages="$packages doxygen"
+    fi
+    sudo dnf install -y $packages
+else  # Debian-like OS
+    packages="$packages g++ libboost-dev"
+    sudo apt-get update -q
+    if [ "$unit_tests" = "true" ]; then
+        packages="$packages libboost-system-dev libboost-filesystem-dev libboost-test-dev"
+    fi
+    if [ "$doxygen" = "true" ]; then
+        packages="$packages doxygen"
+    fi
+    sudo apt-get install -y $packages
+fi
 
 # Download and build the libraries here.
-cd /tmp
+mkdir -p ~/src
+cd ~/src
 
 # Download and install ANN.
-wget -nv https://www.cs.umd.edu/~mount/ANN/Files/1.1.2/ann_1.1.2.tar.gz
+rm -rf ann_1.1.2/
+if [ ! -f "ann_1.1.2.tar.gz" ]; then
+    wget -nv https://www.cs.umd.edu/~mount/ANN/Files/1.1.2/ann_1.1.2.tar.gz
+fi
 tar xzf ann_1.1.2.tar.gz
 cd ann_1.1.2/
 make -j $job_count linux-g++
@@ -45,15 +87,21 @@ cd ..
 # Use the version tagged "g20210303" by FreeBSD ports.
 # C.f. https://cgit.freebsd.org/ports/tree/math/exprtk/Makefile
 exprtk_sha1=ca5c577917646ddba3f71ce6d5dd7d01f351ee80
-wget -nv https://github.com/ArashPartow/exprtk/archive/$exprtk_sha1.zip
-mv $exprtk_sha1.zip exprtk-$exprtk_sha1.zip
+rm -rf exprtk-$exprtk_sha1/
+if [ ! -f "exprtk-$exprtk_sha1.zip" ]; then
+    wget -nv https://github.com/ArashPartow/exprtk/archive/$exprtk_sha1.zip
+    mv $exprtk_sha1.zip exprtk-$exprtk_sha1.zip
+fi
 unzip -q exprtk-$exprtk_sha1.zip
 sudo cp exprtk-$exprtk_sha1/exprtk.hpp /usr/local/include/
 
 # Download, patch and install ScalFMM.
 # Use the tip of the branch maintenance/scalfmm-1.5 as of 2021-09-20.
 scalfmm_sha1=22b9e4f6cf4ea721d71198a71e3f5d2c5ae5e7cc
-wget -nv https://gitlab.inria.fr/solverstack/ScalFMM/-/archive/$scalfmm_sha1/ScalFMM-$scalfmm_sha1.tar.gz
+rm -rf ScalFMM-$scalfmm_sha1/
+if [ ! -f "ScalFMM-$scalfmm_sha1.tar.gz" ]; then
+    wget -nv https://gitlab.inria.fr/solverstack/ScalFMM/-/archive/$scalfmm_sha1/ScalFMM-$scalfmm_sha1.tar.gz
+fi
 tar xzf ScalFMM-$scalfmm_sha1.tar.gz
 cd ScalFMM-$scalfmm_sha1/
 sed -i 's/memcpy/if (nbParticles != 0) memcpy/' Src/Components/FBasicParticleContainer.hpp
@@ -65,7 +113,10 @@ sudo make install
 cd ../..
 
 # Download and install GMM.
-wget -nv http://download-mirror.savannah.gnu.org/releases/getfem/stable/gmm-5.4.tar.gz
+rm -rf gmm-5.4/
+if [ ! -f "gmm-5.4.tar.gz" ]; then
+    wget -nv http://download-mirror.savannah.gnu.org/releases/getfem/stable/gmm-5.4.tar.gz
+fi
 tar xzf gmm-5.4.tar.gz
 cd gmm-5.4/
 ./configure
